requirements.txt
# Microsoft GraphRAG Implementation - Complete Stack
# SuperClaude Wave Orchestration - Phase 2 Complete

# Core GraphRAG dependencies
graphrag==0.3.0
datashaper==0.0.49

# AI & LLM dependencies (Gemini-focused)
google-generativeai==0.8.0
tiktoken==0.7.0
tenacity==8.2.3

# Data processing and analysis
pandas==2.0.3
numpy==1.24.4
networkx==3.1
pyarrow==12.0.1
textstat==0.7.3

# Vector databases and embeddings
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Community detection (optional, fallback available)
python-louvain==0.16

# Async & HTTP
aiohttp==3.9.5
httpx==0.27.0
asyncio-throttle==1.0.2

# FastAPI server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6

# Database support
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.12.1

# Utilities and configuration
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2
rich==13.6.0
devtools==0.12.2

# Scientific computing
scipy==1.11.4
scikit-learn==1.3.2

# Text processing
spacy==3.7.2
nltk==3.8.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.9.1
isort==5.12.0
mypy==1.6.1
flake8==6.1.0

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0